configure: Allow user to specify make tool
configure runs the 'create-dirs' target in Makefile (generated from
Makefile.in). This Makefile requires GNU Make. In FreeBSD systems, the
command 'make' is the BSD make utility and GNU Make is available in the
package management system. When GNU Make is installed in FreeBSD, the
command has the name 'gmake'.

This commit allows the user to specify a make tool with the MAKE
environvent variable. If not specified, 'make' is the default.

Example:
  MAKE=gmake configure