Update Foreign.rst to reflect my learnings (#1195)
* Update Foreign.rst to reflect my learnings

I found the page to be a little unclear in when and why to link to externally generated object files. It is a simple point in hindsight, but moving Linking foreign object files to GHDL to right after Foreign Declarations of Function makes it easier to see the link.
I also added a Hint on how to pass command line arguments to ghdl_main, and wrote that the one code block is Python code (which wasn't immediately obvious to me at least).

* Update Foreign.rst to reflect my learnings: Rev1

Move Linking... back to original spot. 
Better hint, and extra reference to simulation options.

* Add reference-able label to --elab-run

* Update Foreign.rst to reflect my learnings: Rev2

Clean up reference hint.

* Update Foreign.rst to reflect my learnings: Rev3

Further clean reference hints.

* Update Foreign.rst to reflect my learnings: Rev4

Improve hints.

* Move and make -Wl,*.o hint generic, reference it

* Do lables require an empty line to follow??

* Missed 'is'

* Remove promise for example.

* Remove promise again.....

* Give up on referencing hint, just ref "linking...

* Line feed reverted to LF

Co-authored-by: Ross <radonnachie@gmail.com>