vhdl: handle locally static attributes on entity/architecture/configurations
Attributes of entity/architecture/configurations are expected to be locally
static so that they can be referenced from outside (so on the non-instantiated
entity).

But many designs break this assumption.  In relaxed mode, non-locally static
attributes are allowed but now cannot be referenced outside the entity.

Locally static attributes can be referenced from outside.

Fix #1528