Makefile.in: Use $(CURDIR) instead of $(shell pwd)
GNU make will set the variable CURDIR to the current directory, there is
no need to call out to the shell. Also use simple assignment instead of
conditional on not previously set. Otherwise it could get its value from
an envirnoment variable "PWD", which shells may or may not have set, or
which may not be up to date at the time of the call.