Makefile.in: Allow overriding version string even with git
When creating version.tmp, the Makefile.in has provisions for
setting the DESC, REF and HASH variables from environment variables.
However these would be overwritten if it detects it is building from a
git checkout. This commit changes it to always override the default or
autodetected values when the environment variables are available,