Add Allow_Unknown_Subprogram_Body flag
When the sensitivity list is extracted, the subprgram bodies are not
known, which results in an assertion failing in vhdl-nodes.adb. The
solution is enabling a flag so that the assertion is not checked for
the specific case of extracting the sensitivity list. After extracting
the sensitivity list, we must disable the flag.